+++ src/arithimpl/rampimpl.rs
@@ -68,4 +68,4 @@ impl BitManipulation for ramp::Int {
     }
 }
 
-pub type BigInteger = ramp::Int;
+pub type BigInt = ramp::Int;
